refactor: remove impl cell_provider on store
